move file validation
