+++ src/components/Form/Step3/UploadTicket/FileUpload/FileUpload.js
@@ -1,25 +1,37 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 // Import components
 import Icon from 'components/shared/Icon/Icon';
-// Import custom hooks
-import useFileUploadValidation from './useFileUploadValidation';
 // Import contexts
 import { useFormContext } from 'react-hook-form'
 // Import styles
 import s from './FileUpload.module.scss';
 
 const FileUpload = ({name, fieldValidation}) => {
-  // Use custom hook for validating fileUpload inputs
-  const {
-    handleChange,
-    handleBlur,
-    handleFocus,
-    isFileInputFocused,
-    fileName,
-  } = useFileUploadValidation();
   
   const { errors } = useFormContext()
+
+  // Local state for controlling file upload
+  const [isFileInputFocused, setIsFileInputFocused] = useState(false); // This is used to emulate the input focus class on the label
+  const [fileName, setFileName] = useState('Upload photo'); // Used to change the name of the input/label button to the users file name
+
+
+  const handleChange = (e) => {
+    const file = e.target.files[0];
+
+    // If a file exists (user hasn't clicked cancel button or something)
+    if (file) {
+      setFileName(file.name); // Set file name that the user has chosen (this will display in our label)
+    }
+  };
+
+  // HandleFocus (when user joins input)
+  const handleFocus = () => {
+    setIsFileInputFocused(true); // Set input to focus
+  };
+
+  // Handleblur (when user leaves input), set input to unfocus
+  const handleBlur = () => setIsFileInputFocused(false);
 
   return (
     <div className={`wmnds-fe-group ${errors[name] ? 'wmnds-fe-group--error' : ''}`}>
